Prepare for early merge
